+++ cli/get_stats.py
@@ -6,8 +6,8 @@ import click
 from dateutil.relativedelta import relativedelta
 import pandas as pd
 
-import dlstats
-from xdevrepos import XdevRepos
+from ..utilities import download_stats
+from ..utilities.xdev_repos import XdevRepos
 
 
 def update_dataframe(df1, df2):
@@ -70,9 +70,9 @@ def main(sources, last, all):
 
     for source in sources:
         packages = repos.packages(source=source)
-        method = getattr(dlstats, f'get_{source}_stats')
+        method = getattr(download_stats, f'get_{source}_stats')
 
-        statsfile = Path(f'{source}stats.csv')
+        statsfile = '../data' / Path(f'{source}_stats.csv')
         if statsfile.exists():
             print(f'Reading {source} statistics file...')
             df = pd.read_csv(statsfile)
